Add: products integration tests

+++ src/http/routes/products/delete-product.ts
@@ -1,3 +1,5 @@
+import { retrieveProduct } from '@/app/functions/products/retrieveProduct'
+import { retrieveRestaurant } from '@/app/functions/retrieveRestaurant'
 import { db } from '@/database/client'
 import { products } from '@/database/schemas/products'
 import { eq } from 'drizzle-orm'
@@ -9,6 +11,24 @@ export const deleteProductRoute = new Elysia().use(betterAuthPluggin).delete(
   async ({ status, params, user }) => {
     const { productId } = params
 
+    const { foundedProduct } = await retrieveProduct(productId)
+
+    if (!foundedProduct) {
+      return status(404, {
+        message: 'Product not found.',
+      })
+    }
+
+    const { restaurant: productOwner } = await retrieveRestaurant(
+      foundedProduct.producer
+    )
+
+    if (!productOwner || user.id !== productOwner.owner) {
+      return status(401, {
+        message: 'You are not the producer of this product.',
+      })
+    }
+
     await db.delete(products).where(eq(products.id, productId))
 
     return status(204, void 0)
@@ -17,8 +37,8 @@ export const deleteProductRoute = new Elysia().use(betterAuthPluggin).delete(
     auth: true,
     detail: {
       tags: ['Products'],
-      description: 'List all Products',
-      operationId: 'listProducts',
+      description: 'Delete a Product',
+      operationId: 'deleteProduct',
       security: [
         {
           bearerAuth: [],
@@ -33,6 +53,16 @@ export const deleteProductRoute = new Elysia().use(betterAuthPluggin).delete(
     }),
     response: {
       204: t.Void(),
+      401: t.Object({
+        message: t.String({
+          examples: ['Unauthorized.'],
+        }),
+      }),
+      404: t.Object({
+        message: t.String({
+          examples: ['Product not found.'],
+        }),
+      }),
       500: t.Object({
         message: t.String({
           examples: ['Internal Server Error.'],
